Improve observer readiness, recovery, and runtime documentation
