+++ docs/architecture/runtime-support.md
@@ -1,0 +1,49 @@
+# Runtime support and recovery
+
+<!-- Generated by scripts/generate_runtime_support.py; edit its sources, then regenerate. -->
+
+This table describes implemented eligibility and assigned CI coverage. It does
+not certify an installation or a later commit. A command can execute successfully
+without qualifying for automatic splitting or observation-based reuse.
+
+| Capability | Bounded profile | Result when outside the profile |
+| --- | --- | --- |
+| Python inventory / splitting | CPython >= 3.10.0, < 3.15.0; supported unittest / pytest collection | Keep the parent when completeness cannot be verified |
+| Vitest inventory / splitting | 5.0.0; supported static configuration | Keep the parent |
+| Jest inventory / splitting | package 30.5.1, CLI 30.5.0; supported static configuration | Keep the parent |
+| Complete native Python observation | CPython 3.12.3–3.12.14; platform companion and complete bound inputs required | Execute the check; no complete-observation authority |
+| pytest observation CI | 8.4.2, 9.1.1; constrained options, plugins and input behavior | Execute when observation is incomplete |
+| Conditional JavaScript observation | Linux Node v22.23.2; eligible Node/Vitest/Jest execution and observed-input receipts | Execute affected children; no claim of complete JS capture |
+
+Splitting permission and reuse permission are separate. An invalid receipt runs
+that child; only an unverifiable complete partition requires the parent. Ordinary
+execution and exact-receipt / owner-policy routes remain available under their
+own checks. See [automatic sharding](../../skills/click/references/automatic-sharding-setup.md).
+
+Native collection uses Linux strace 6.8, macOS privileged fs_usage, or Windows
+ETW (logman and tracerpt), plus a matching compiler, headers and companion.
+Click neither installs these prerequisites nor elevates privileges. Python
+deterministic input capture requires PYTHONHASHSEED=0 and
+PYTHONDONTWRITEBYTECODE=1; explicit conflicting user values are preserved.
+See [native requirements](../../skills/click/references/authoritative-observer-v2.md).
+
+Use `click-gate observer status` or `click-gate verification status` for recorded
+preparation reasons and next actions. These views never run a collector and
+always report `reuse_authorized: false`. A prepared runtime is not a passing
+receipt. The dashboard separates preparation from the last recorded per-check
+decision and labels conditional JS reuse as input-completeness-unproven.
+
+Missing tools: install the documented prerequisite using your normal host setup.
+Unsupported runtime: use a listed runtime or keep normal execution. Insufficient
+permissions: review host permissions; status never requests automatic elevation.
+Automatic preparation retries after relevant environment, tool, owner-policy or
+runtime changes; `click-gate observer auto` permits an explicit retry. Unchanged
+failed contexts do not repeatedly prepare. Recovery never changes requested test
+options or turns diagnostic data into reuse authority. Windows collector errors
+emit bounded stage codes, never raw command output or private paths.
+
+Conditional JS reuse requires matching separately attested observations from
+requested executions. Known clock/random/shared-memory activity, unsupported
+process coverage and input gaps remain ineligible. See [the exact conditional
+scope](node-runtime-observation.md). No extra test execution is launched merely
+to learn a conditional baseline.
