+++ hooks/dashboard/locales/ko.json
@@ -637,5 +637,38 @@
   "Observer: 런타임 진단": "Observer: 런타임 진단",
   "시간·난수·공유 메모리 사용을 진단합니다. 이 관찰로 재사용을 허용하지 않습니다.": "시간·난수·공유 메모리 사용을 진단합니다. 이 관찰로 재사용을 허용하지 않습니다.",
   "조건부 재사용: 관찰된 파일·환경이 유지됩니다. 미관찰 입력은 보증하지 않으며 입력 완전성은 미검증입니다.": "조건부 재사용: 관찰된 파일·환경이 유지됩니다. 미관찰 입력은 보증하지 않으며 입력 완전성은 미검증입니다.",
-  "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증": "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증"
+  "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증": "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증",
+  "준비 상태는 재사용 승인이 아닙니다. 조건부 JS 재사용은 관찰된 입력만 확인하며 완전성은 입증되지 않았습니다.": "준비 상태는 재사용 승인이 아닙니다. 조건부 JS 재사용은 관찰된 입력만 확인하며 완전성은 입증되지 않았습니다.",
+  "검사별 재실행 이유와 다음 동작": "검사별 재실행 이유와 다음 동작",
+  "다음 검증 요청에서 지원되는 입력 관찰을 준비합니다.": "다음 검증 요청에서 지원되는 입력 관찰을 준비합니다.",
+  "수집기 준비 기록이 있습니다. 검증 요청마다 입력과 영수증을 다시 확인합니다.": "수집기 준비 기록이 있습니다. 검증 요청마다 입력과 영수증을 다시 확인합니다.",
+  "입력 관찰을 껐습니다. 기존 영수증과 저장소 정책 재사용은 계속 사용할 수 있습니다.": "입력 관찰을 껐습니다. 기존 영수증과 저장소 정책 재사용은 계속 사용할 수 있습니다.",
+  "현재 모드는 진단 정보만 수집합니다.": "현재 모드는 진단 정보만 수집합니다.",
+  "저장소 소유자가 선택한 입력 정책을 사용합니다.": "저장소 소유자가 선택한 입력 정책을 사용합니다.",
+  "현재 Python 환경은 자동 관찰 조건과 다릅니다. 원래 검사는 그대로 실행합니다.": "현재 Python 환경은 자동 관찰 조건과 다릅니다. 원래 검사는 그대로 실행합니다.",
+  "이 Python 실행 환경은 자동 관찰 지원 범위 밖입니다.": "이 Python 실행 환경은 자동 관찰 지원 범위 밖입니다.",
+  "수집 도구 또는 빌드 도구를 사용할 수 없습니다.": "수집 도구 또는 빌드 도구를 사용할 수 없습니다.",
+  "현재 권한으로 수집기를 사용할 수 없습니다.": "현재 권한으로 수집기를 사용할 수 없습니다.",
+  "입력 관찰을 준비하지 못했습니다. 일반 검증은 사용할 수 있습니다.": "입력 관찰을 준비하지 못했습니다. 일반 검증은 사용할 수 있습니다.",
+  "평소 검증 명령을 실행하세요. 관찰 때문에 추가 실행하지 않습니다.": "평소 검증 명령을 실행하세요. 관찰 때문에 추가 실행하지 않습니다.",
+  "기존 입력 정책을 유지하고 평소처럼 검증하세요.": "기존 입력 정책을 유지하고 평소처럼 검증하세요.",
+  "자동 관찰에는 PYTHONHASHSEED=0, PYTHONDONTWRITEBYTECODE=1이 필요합니다. 테스트 의미에 맞는 경우에만 환경을 선택하세요.": "자동 관찰에는 PYTHONHASHSEED=0, PYTHONDONTWRITEBYTECODE=1이 필요합니다. 테스트 의미에 맞는 경우에만 환경을 선택하세요.",
+  "지원 범위는 CPython 3.12.3–3.12.14입니다. 다른 버전도 일반 검증은 계속 실행합니다.": "지원 범위는 CPython 3.12.3–3.12.14입니다. 다른 버전도 일반 검증은 계속 실행합니다.",
+  "Linux는 strace 6.8와 C 컴파일러, macOS는 fs_usage와 C 컴파일러, Windows는 ETW 도구와 MSVC가 필요합니다.": "Linux는 strace 6.8와 C 컴파일러, macOS는 fs_usage와 C 컴파일러, Windows는 ETW 도구와 MSVC가 필요합니다.",
+  "권한을 자동으로 올리지 않습니다. 관찰 없이 검증하거나 권한 있는 환경에서 수동으로 준비하세요.": "권한을 자동으로 올리지 않습니다. 관찰 없이 검증하거나 권한 있는 환경에서 수동으로 준비하세요.",
+  "준비 조건이 바뀌면 자동으로 재확인합니다. 수동 재시도: click-gate observer auto": "준비 조건이 바뀌면 자동으로 재확인합니다. 수동 재시도: click-gate observer auto",
+  "이 선택을 유지합니다. 모드 확인: click-gate observer status": "이 선택을 유지합니다. 모드 확인: click-gate observer status",
+  "조건부 JS 재사용의 입력 완전성은 입증되지 않았습니다. 알려진 동적 입력은 다시 실행합니다.": "조건부 JS 재사용의 입력 완전성은 입증되지 않았습니다. 알려진 동적 입력은 다시 실행합니다.",
+  "영향받은 검사를 다시 실행하세요. 형제 결과는 각자의 근거로 재판정합니다.": "영향받은 검사를 다시 실행하세요. 형제 결과는 각자의 근거로 재판정합니다.",
+  "최근 계획에 기록된 재사용입니다. 현재 결과 확인: click-gate status": "최근 계획에 기록된 재사용입니다. 현재 결과 확인: click-gate status",
+  "조건부 재사용": "조건부 재사용",
+  "입력 변경": "입력 변경",
+  "실행 조건 변경": "실행 조건 변경",
+  "실제 출력 필요": "실제 출력 필요",
+  "입력 근거 부족": "입력 근거 부족",
+  "재사용 계획 기록": "재사용 계획 기록",
+  "검증 필요": "검증 필요",
+  "이전 형식의 기록입니다. 최신 검증 요청에서 준비 상태를 확인하세요.": "이전 형식의 기록입니다. 최신 검증 요청에서 준비 상태를 확인하세요.",
+  "입력 관찰 상태를 확인하세요.": "입력 관찰 상태를 확인하세요.",
+  "아직 검사별 실행 계획이 없습니다.": "아직 검사별 실행 계획이 없습니다."
 }
