+++ hooks/dashboard/locales/zh-CN.json
@@ -637,5 +637,38 @@
   "Observer: 런타임 진단": "Observer：运行时诊断",
   "시간·난수·공유 메모리 사용을 진단합니다. 이 관찰로 재사용을 허용하지 않습니다.": "诊断时间、随机数和共享内存使用。这些观测不授予复用权限。",
   "조건부 재사용: 관찰된 파일·환경이 유지됩니다. 미관찰 입력은 보증하지 않으며 입력 완전성은 미검증입니다.": "条件复用：观测到的文件和环境未变。未观测输入不在保证范围内，输入完整性尚未验证。",
-  "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증": "条件 JS 复用：{0} 项 · 仅限观测输入 / 完整性未获证明"
+  "조건부 JS 재사용 {0}개 · 관찰 범위 기반 / 입력 완전성 미보증": "条件 JS 复用：{0} 项 · 仅限观测输入 / 完整性未获证明",
+  "준비 상태는 재사용 승인이 아닙니다. 조건부 JS 재사용은 관찰된 입력만 확인하며 완전성은 입증되지 않았습니다.": "准备状态不授予复用权限。JS 条件复用仅检查已观测输入，完整性尚未证明。",
+  "검사별 재실행 이유와 다음 동작": "各项检查重新执行的原因与下一步",
+  "다음 검증 요청에서 지원되는 입력 관찰을 준비합니다.": "将在下次验证请求时准备支持的输入采集。",
+  "수집기 준비 기록이 있습니다. 검증 요청마다 입력과 영수증을 다시 확인합니다.": "已有采集器准备记录。每次验证请求都会重新检查输入和凭据。",
+  "입력 관찰을 껐습니다. 기존 영수증과 저장소 정책 재사용은 계속 사용할 수 있습니다.": "输入观测已关闭。已有凭据和仓库策略复用仍可使用。",
+  "현재 모드는 진단 정보만 수집합니다.": "当前模式仅采集诊断信息。",
+  "저장소 소유자가 선택한 입력 정책을 사용합니다.": "使用仓库所有者选择的输入策略。",
+  "현재 Python 환경은 자동 관찰 조건과 다릅니다. 원래 검사는 그대로 실행합니다.": "当前 Python 环境不满足自动观测条件，原始检查仍正常执行。",
+  "이 Python 실행 환경은 자동 관찰 지원 범위 밖입니다.": "此 Python 运行环境不在自动观测支持范围内。",
+  "수집 도구 또는 빌드 도구를 사용할 수 없습니다.": "采集工具或构建工具不可用。",
+  "현재 권한으로 수집기를 사용할 수 없습니다.": "当前权限无法运行采集器。",
+  "입력 관찰을 준비하지 못했습니다. 일반 검증은 사용할 수 있습니다.": "输入观测准备失败，普通验证仍可使用。",
+  "평소 검증 명령을 실행하세요. 관찰 때문에 추가 실행하지 않습니다.": "请运行平常的验证命令。不会为观测额外执行测试。",
+  "기존 입력 정책을 유지하고 평소처럼 검증하세요.": "保留现有输入策略并正常验证。",
+  "자동 관찰에는 PYTHONHASHSEED=0, PYTHONDONTWRITEBYTECODE=1이 필요합니다. 테스트 의미에 맞는 경우에만 환경을 선택하세요.": "自动观测要求 PYTHONHASHSEED=0 和 PYTHONDONTWRITEBYTECODE=1。仅在符合测试语义时选择此环境。",
+  "지원 범위는 CPython 3.12.3–3.12.14입니다. 다른 버전도 일반 검증은 계속 실행합니다.": "观测支持 CPython 3.12.3–3.12.14，其他版本仍可执行普通验证。",
+  "Linux는 strace 6.8와 C 컴파일러, macOS는 fs_usage와 C 컴파일러, Windows는 ETW 도구와 MSVC가 필요합니다.": "Linux 需要 strace 6.8 和 C 编译器；macOS 需要 fs_usage 和 C 编译器；Windows 需要 ETW 工具和 MSVC。",
+  "권한을 자동으로 올리지 않습니다. 관찰 없이 검증하거나 권한 있는 환경에서 수동으로 준비하세요.": "不会自动提升权限。可不启用观测进行验证，或在有权限的环境中手动准备。",
+  "준비 조건이 바뀌면 자동으로 재확인합니다. 수동 재시도: click-gate observer auto": "准备条件变化后会自动重新检查。手动重试：click-gate observer auto",
+  "이 선택을 유지합니다. 모드 확인: click-gate observer status": "保留当前选择。查看模式：click-gate observer status",
+  "조건부 JS 재사용의 입력 완전성은 입증되지 않았습니다. 알려진 동적 입력은 다시 실행합니다.": "JS 条件复用不证明输入完整性，已知动态输入需要重新执行。",
+  "영향받은 검사를 다시 실행하세요. 형제 결과는 각자의 근거로 재판정합니다.": "请重新执行受影响的检查，兄弟检查按各自证据重新判定。",
+  "최근 계획에 기록된 재사용입니다. 현재 결과 확인: click-gate status": "这是最近计划记录的复用。查看当前结果：click-gate status",
+  "조건부 재사용": "条件复用",
+  "입력 변경": "输入已变化",
+  "실행 조건 변경": "执行条件已变化",
+  "실제 출력 필요": "需要实际输出",
+  "입력 근거 부족": "输入证据不足",
+  "재사용 계획 기록": "已记录复用计划",
+  "검증 필요": "需要验证",
+  "이전 형식의 기록입니다. 최신 검증 요청에서 준비 상태를 확인하세요.": "此记录使用旧格式，请在新的验证请求后查看准备状态。",
+  "입력 관찰 상태를 확인하세요.": "请查看输入观测状态。",
+  "아직 검사별 실행 계획이 없습니다.": "尚未记录各项检查的执行计划。"
 }
