+++ hooks/dashboard/styles.css
@@ -277,3 +277,8 @@ footer { display:flex; justify-content:space-between; gap:12px; color:var(--mute
 @media(max-width:760px) { body{padding:9px} .app-shell{grid-template-columns:1fr;border-radius:17px} .sidebar{position:relative;padding:14px 13px 9px;border-right:0;border-bottom:1px solid #d1e6e6} .brand{font-size:29px;margin:0 0 16px;padding:0;gap:8px} .mark{width:29px;height:33px;font-size:25px;border-radius:10px} .mark span{font-size:13px;top:-4px} .sidebar nav{display:flex;gap:4px;overflow:auto} .nav-link{white-space:nowrap;min-height:39px;padding:9px 10px;font-size:11px;gap:7px;border-radius:8px} .nav-link .ui-icon{width:16px;height:16px} .nav-caption,.sidebar-note{display:none} .sidebar-bottom{display:none} .language-picker{width:100%;min-height:44px;gap:9px} .language-picker select{margin-left:auto;min-width:128px;font-size:13px} main{padding:18px 12px 12px} .topbar{gap:10px;margin-bottom:12px} .topbar h1{font-size:25px;max-width:100%;line-height:1.4} .topbar .eyebrow{font-size:8px} .page-subtitle{font-size:11px;margin-top:6px} .top-actions{align-self:stretch;justify-content:flex-end;gap:12px} .top-actions .live{margin-right:auto;font-size:10px} .icon-button{width:32px;height:32px;font-size:23px} .primary-link{padding:8px 12px;font-size:11px;border-radius:8px} .panel{padding:16px;border-radius:14px} .task-summary{grid-template-columns:34px minmax(0,1fr);gap:0 9px;padding:12px;margin-bottom:12px} .task-icon{width:34px;height:38px;border-radius:9px} .task-icon .ui-icon{width:21px;height:21px} .task-identity h2{font-size:13px} .task-identity .eyebrow{font-size:9px} #taskline{font-size:10px} .task-meta{grid-column:1/-1;grid-row:3;display:flex;justify-content:space-between;gap:10px;align-items:center;margin-top:10px;padding-top:8px;border-top:1px solid #dbe9ec} .task-meta time{font-size:9px} #approvalState{font-size:9px} .task-summary details[open]{grid-column:1/-1} .readiness-panel{padding:13px;margin-bottom:12px}.readiness-copy h2{font-size:14px}.readiness-items{grid-template-columns:repeat(5,108px)} .panel-title{margin-bottom:13px} .panel-title h2{font-size:18px} .panel-subtitle{font-size:11px} .impact-grid{gap:12px} .headline-metrics{grid-template-columns:1fr;gap:10px} .headline-card{padding:13px 14px} .metric-heading h2{font-size:15px} .hero-number{margin:8px 0 5px} .hero-value,:lang(en) .hero-value{font-size:34px} .hero-status{font-size:10px} .savings-breakdown{margin-top:13px} .breakdown-row{font-size:11px} .task-effect-value{font-size:19px} .truth-note{font-size:9px} .metrics{grid-template-columns:repeat(2,minmax(0,1fr));gap:12px 0;margin-bottom:14px} .metrics article:nth-child(3){border:0;padding-left:0} .metrics strong,.metric-link{font-size:28px} .baseline-row{display:block} .baseline-row .row-label{margin-bottom:5px} .group-blocks{grid-template-columns:repeat(auto-fit,minmax(31px,1fr))} .comparison-values{grid-template-columns:1fr;gap:8px} .comparison-values p{grid-template-columns:1fr 1fr;padding:0;border:0;align-items:center} .comparison-values strong{text-align:right;font-size:15px} .axis>div{grid-template-columns:87px minmax(0,1fr);gap:7px} .axis>div>span{font-size:8px} .track{height:12px} .work-grid{gap:12px;margin-top:12px} .source-table{min-width:465px} .source-scroll{max-height:235px} #filterStatus{font-size:8px} .source-filters button{font-size:10px;padding:4px 8px} .dashboard-note{font-size:11px;gap:9px;padding:10px} .dashboard-note .ui-icon{width:19px} .explanation{padding:12px 15px} .metric-split{grid-template-columns:1fr;gap:12px} .measurement-details{padding:13px 15px} .statline{font-size:10px} #contractPromises{display:block} .exports textarea{font-size:11px} .export-actions{gap:7px} .export-actions button{min-height:38px;font-size:10px} footer{flex-wrap:wrap;font-size:9px} }
 @media(prefers-reduced-motion:reduce) { html{scroll-behavior:auto} *,*:before,*:after{animation:none!important;transition:none!important} }
 @media(max-width:760px) { .sidebar{top:auto;height:auto;overflow:visible} }
+
+.observer-readiness { grid-column: 1 / -1; min-width: 0; padding-top: 12px; border-top: 1px solid #d5e5ed; }
+.observer-readiness p { margin: 6px 0; overflow-wrap: anywhere; }
+.readiness-checks { max-height: 240px; overflow: auto; padding-left: 24px; }
+.readiness-checks li { margin: 8px 0; overflow-wrap: anywhere; }
