+++ scripts/generate_runtime_support.py
@@ -1,0 +1,106 @@
+#!/usr/bin/env python3
+"""Generate the capability table from literal source constants; never probe tools."""
+from __future__ import annotations
+
+import argparse
+import ast
+from pathlib import Path
+import re
+
+ROOT = Path(__file__).resolve().parents[1]
+TARGET = "docs/architecture/runtime-support.md"
+
+
+def constant(root: Path, module: str, name: str):
+    tree = ast.parse((root / "hooks" / (module + ".py")).read_text(encoding="utf-8"))
+    for statement in tree.body:
+        if isinstance(statement, ast.Assign) and any(isinstance(target, ast.Name) and target.id == name for target in statement.targets):
+            return ast.literal_eval(statement.value)
+    raise ValueError(f"Missing literal capability: {module}.{name}")
+
+
+def render(root: Path = ROOT) -> str:
+    def version(module, name):
+        value = constant(root, module, name)
+        return ".".join(map(str, value)) if isinstance(value, tuple) else value
+    minimum = version("click_observer_profiles", "MIN_NATIVE_CPYTHON")
+    maximum = version("click_observer_profiles", "MAX_NATIVE_CPYTHON")
+    inventory_min = version("click_collector_runtime", "MIN_CPYTHON")
+    inventory_max = version("click_collector_runtime", "MAX_CPYTHON")
+    node = version("click_node_observer", "VERSION")
+    vitest = version("click_test_inventory", "VITEST_VERSION")
+    jest = version("click_test_inventory", "JEST_VERSION")
+    jest_cli = version("click_test_inventory", "JEST_CLI_VERSION")
+    workflow = (root / ".github/workflows/ci.yml").read_text(encoding="utf-8")
+    matrices = re.findall(r"^\s+pytest: (\[[^\n]+\])$", workflow, re.MULTILINE)
+    if len(matrices) != 1:
+        raise ValueError("Expected one pinned pytest observation matrix")
+    pytest = ", ".join(ast.literal_eval(matrices[0]))
+    return f"""# Runtime support and recovery
+
+<!-- Generated by scripts/generate_runtime_support.py; edit its sources, then regenerate. -->
+
+This table describes implemented eligibility and assigned CI coverage. It does
+not certify an installation or a later commit. A command can execute successfully
+without qualifying for automatic splitting or observation-based reuse.
+
+| Capability | Bounded profile | Result when outside the profile |
+| --- | --- | --- |
+| Python inventory / splitting | CPython >= {inventory_min}, < {inventory_max}; supported unittest / pytest collection | Keep the parent when completeness cannot be verified |
+| Vitest inventory / splitting | {vitest}; supported static configuration | Keep the parent |
+| Jest inventory / splitting | package {jest}, CLI {jest_cli}; supported static configuration | Keep the parent |
+| Complete native Python observation | CPython {minimum}–{maximum}; platform companion and complete bound inputs required | Execute the check; no complete-observation authority |
+| pytest observation CI | {pytest}; constrained options, plugins and input behavior | Execute when observation is incomplete |
+| Conditional JavaScript observation | Linux Node {node}; eligible Node/Vitest/Jest execution and observed-input receipts | Execute affected children; no claim of complete JS capture |
+
+Splitting permission and reuse permission are separate. An invalid receipt runs
+that child; only an unverifiable complete partition requires the parent. Ordinary
+execution and exact-receipt / owner-policy routes remain available under their
+own checks. See [automatic sharding](../../skills/click/references/automatic-sharding-setup.md).
+
+Native collection uses Linux strace 6.8, macOS privileged fs_usage, or Windows
+ETW (logman and tracerpt), plus a matching compiler, headers and companion.
+Click neither installs these prerequisites nor elevates privileges. Python
+deterministic input capture requires PYTHONHASHSEED=0 and
+PYTHONDONTWRITEBYTECODE=1; explicit conflicting user values are preserved.
+See [native requirements](../../skills/click/references/authoritative-observer-v2.md).
+
+Use `click-gate observer status` or `click-gate verification status` for recorded
+preparation reasons and next actions. These views never run a collector and
+always report `reuse_authorized: false`. A prepared runtime is not a passing
+receipt. The dashboard separates preparation from the last recorded per-check
+decision and labels conditional JS reuse as input-completeness-unproven.
+
+Missing tools: install the documented prerequisite using your normal host setup.
+Unsupported runtime: use a listed runtime or keep normal execution. Insufficient
+permissions: review host permissions; status never requests automatic elevation.
+Automatic preparation retries after relevant environment, tool, owner-policy or
+runtime changes; `click-gate observer auto` permits an explicit retry. Unchanged
+failed contexts do not repeatedly prepare. Recovery never changes requested test
+options or turns diagnostic data into reuse authority. Windows collector errors
+emit bounded stage codes, never raw command output or private paths.
+
+Conditional JS reuse requires matching separately attested observations from
+requested executions. Known clock/random/shared-memory activity, unsupported
+process coverage and input gaps remain ineligible. See [the exact conditional
+scope](node-runtime-observation.md). No extra test execution is launched merely
+to learn a conditional baseline.
+"""
+
+
+def main() -> int:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--check", action="store_true")
+    args = parser.parse_args()
+    path = ROOT / TARGET
+    expected = render()
+    if args.check:
+        if not path.exists() or path.read_text(encoding="utf-8") != expected:
+            parser.exit(1, f"Stale {TARGET}; run scripts/generate_runtime_support.py\n")
+    else:
+        path.write_text(expected, encoding="utf-8")
+    return 0
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
